// types.ts
export interface Trend {
    name: string;
    prompt: string;
    bg: string;
    icon: string;
    type?: 'descriptive';
}